requirements.txt
beautifulsoup4
flask
requests
psycopg2-binary